Added styles to form layout, legend, text area and spinner.
I put a border radius on the layout of the form, i also added a background color. Then i added a background color to the legend and a font color. Then i added a height and width to the text area styles and lastly i added a webkit flex to the spinner styles.